requirements.txt
dask
pandas
boto3
psycopg2-binary
cassandra-driver
pyspark==3.5.5
numpy
requests
minio
apache-airflow-providers-apache-spark
argparse